Handle `Result` on clipboard `set_contents`
Change the colour of the selection arrow to red if there's an error and
green if the clipboard was written to successfully.

The colour only lasts for 100 milliseconds because that's the refresh
rate of the loop. This seems to be a nice way to briefly offer a visual
notification and automatically return the interface to normal right
afterward.