Parse options from STDIN
Read STDIN and get all lines that start with "e: ", "u: ", and "p: ". If
none are found, exit.

These lines get added to the `options` vector which then gets displayed
in the TUI. Since the contents of the vector are different from our
hard-coded one, use a borrow on the clipboard write call.