Call `pass show` with the given argument and run Passextract on result
Temporarily comment out code that deals with STDIN to get this feature
working. It complicated things to leave it in right now.

Given an argument that isn't "-" (STDIN), we pass that argument to `pass
show`. Users correctly get a password prompt from `pass`, and the output
from the command gets sent to the Passextract TUI for display in the
menu and copying.