ClipboardStore proof of concept
The new type `ClipboardStore` wraps `ClipboardContext`, allowing us to
save the contents of the clipboard before resetting it. We will then be
able to reset the clipboard to this original value.

This commit includes some old code from when I was originally thinking
about this problem before I decided to make a separate module for it.
That code should be deleted.

One caveat to think about in the future is how to reset the clipboard to
the original value after a period of time. In other words, how to pass
the original contents value to a spawned child process.

Also, we need to make sure that we're only setting `last` once.
Subsequent calls should not change the `last` value. And now that I
think about that, "last" is probably the wrong name for this. It should
be something like "original".