Make password line invisible
Instead of rendering the password string, hide it while allowing it to
be copied. It is identified by the "p: " prefix, but nothing follows
that.

This allows you to run `passextract` without having to worry about
others peering over your shoulder or about screen capture or recording
hacks.

I'll be adding a command line flag to force you to opt in to this
feature, while keeping the old functionality the same.