Use `-i` argument to hide passwords
Connect the `-i` argument flag to the renderer. If the arg was passed,
hide passwords, otherwise show them.

Change `args.first()` to `args.get(1)` because the actual first element
is the command executable.

Not happy with the condition I'm using here to grab the password file or
the STDIN `-`. It doesn't read well at all. It all does work, but this
part I want to rewrite to make it easier to read.