When `-i`, replace password with '*'s
Previously, when the "invisible" flag was activated, we wouldn't print
anything except the "p: " prefix on a password line. This made it
difficult to verify the password, and if you for some reason had
multiple passwords, you wouldn't know which one to pick unless you knew
the order.

This at least gives you a visual cue about your password.