Spawn `pass` in order to support GnuPG 2 and Pinentry
Previously, Passextract didn't work with GnuPG 2 because I was
immediately collecting the output of `pass`.

With GnuPG 1.4, the password prompt was like a conventional `read`-style
password prompt. However, in GnuPG 2, the prompt is replaced with
Pinentry. Pinentry interferes with capturing the output from Pass, such
that Passextract exits immediately, without even prompting for a
password.

Instead of using `Command.output`, use `Command.spawn`. This allows
GnuPG to work correctly and display a Pinentry prompt. We can then
capture the standard output from the child process and parse it for the
Passextract interface.